changed URL names
