+++ src/main/java/mate/academy/internetshop/controllers/InjectDataController.java
@@ -12,7 +12,7 @@ import mate.academy.internetshop.model.User;
 import mate.academy.internetshop.service.ProductService;
 import mate.academy.internetshop.service.UserService;
 
-@WebServlet("/injectData")
+@WebServlet("/inject/data")
 public class InjectDataController extends HttpServlet {
     private static final Injector INJECTOR = Injector.getInstance("mate.academy.internetshop");
     private UserService userService
