+++ src/main/java/mate/academy/internetshop/controllers/OrderDeleteController.java
@@ -9,16 +9,16 @@ import javax.servlet.http.HttpServletResponse;
 import mate.academy.internetshop.lib.Injector;
 import mate.academy.internetshop.service.OrderService;
 
-@WebServlet("/deleteOrder")
-public class DeleteOrderController extends HttpServlet {
+@WebServlet("/order/delete")
+public class OrderDeleteController extends HttpServlet {
     private static final Injector INJECTOR = Injector.getInstance("mate.academy.internetshop");
     private OrderService orderService = (OrderService)INJECTOR.getInstance(OrderService.class);
 
     @Override
     protected void doGet(HttpServletRequest req, HttpServletResponse resp)
             throws ServletException, IOException {
-        Long orderId = Long.parseLong(req.getParameter("orderId"));
+        Long orderId = Long.parseLong(req.getParameter("order_id"));
         orderService.delete(orderId);
-        resp.sendRedirect(req.getContextPath() + "/getUserOrders");
+        resp.sendRedirect(req.getContextPath() + "/user/orders");
     }
 }
