+++ src/main/java/mate/academy/internetshop/controllers/UserGetAllController.java
@@ -1,23 +1,28 @@
 package mate.academy.internetshop.controllers;
 
 import java.io.IOException;
+import java.util.List;
+import javax.servlet.ServletException;
 import javax.servlet.annotation.WebServlet;
 import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import mate.academy.internetshop.lib.Injector;
+import mate.academy.internetshop.model.User;
 import mate.academy.internetshop.service.UserService;
 
-@WebServlet("/user/delete")
-public class UserDeleteController extends HttpServlet {
+@WebServlet("/users/all")
+public class UserGetAllController extends HttpServlet {
     private static final Injector INJECTOR = Injector.getInstance("mate.academy.internetshop");
     private UserService userService
             = (UserService) INJECTOR.getInstance(UserService.class);
 
     @Override
     protected void doGet(HttpServletRequest req, HttpServletResponse resp)
-            throws IOException {
-        userService.delete(Long.parseLong(req.getParameter("user_id")));
-        resp.sendRedirect(req.getContextPath() + "/users/all");
+            throws ServletException, IOException {
+
+        List<User> userList = userService.getAll();
+        req.setAttribute("users", userList);
+        req.getRequestDispatcher("/WEB-INF/jsp/allUsers.jsp").forward(req, resp);
     }
 }
