+++ src/main/java/mate/academy/internetshop/controllers/UserGetOrderController.java
@@ -7,22 +7,23 @@ import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import mate.academy.internetshop.lib.Injector;
-import mate.academy.internetshop.model.User;
+import mate.academy.internetshop.model.Order;
 import mate.academy.internetshop.service.OrderService;
-import mate.academy.internetshop.service.UserService;
 
-@WebServlet("/getUserOrders")
-public class GetUserOrdersController extends HttpServlet {
+@WebServlet("/user/order")
+public class UserGetOrderController extends HttpServlet {
     private static final Injector INJECTOR = Injector.getInstance("mate.academy.internetshop");
-    private UserService userService = (UserService) INJECTOR.getInstance(UserService.class);
     private OrderService orderService = (OrderService) INJECTOR.getInstance(OrderService.class);
 
     @Override
     protected void doGet(HttpServletRequest req, HttpServletResponse resp)
             throws ServletException, IOException {
-        Long userId = (Long) req.getSession().getAttribute("user_id");
-        User user = userService.get(userId);
-        req.setAttribute("order", orderService.getUserOrders(user));
-        req.getRequestDispatcher("/WEB-INF/jsp/userOrders.jsp").forward(req, resp);
+        Long orderId = Long.parseLong(req.getParameter("order_id"));
+        Order order = orderService.get(orderId);
+        req.setAttribute("id", orderId);
+        req.setAttribute("date", order.getDateTime());
+        req.setAttribute("products", order.getProducts());
+        req.setAttribute("price", orderService.getPrice(order));
+        req.getRequestDispatcher("/WEB-INF/jsp/userOrder.jsp").forward(req, resp);
     }
 }
