+++ src/main/java/mate/academy/internetshop/controllers/UserGetOrdersController.java
@@ -1,7 +1,6 @@
 package mate.academy.internetshop.controllers;
 
 import java.io.IOException;
-import java.util.List;
 import javax.servlet.ServletException;
 import javax.servlet.annotation.WebServlet;
 import javax.servlet.http.HttpServlet;
@@ -9,20 +8,21 @@ import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import mate.academy.internetshop.lib.Injector;
 import mate.academy.internetshop.model.User;
+import mate.academy.internetshop.service.OrderService;
 import mate.academy.internetshop.service.UserService;
 
-@WebServlet("/allUsers")
-public class GetAllUsersController extends HttpServlet {
+@WebServlet("/user/orders")
+public class UserGetOrdersController extends HttpServlet {
     private static final Injector INJECTOR = Injector.getInstance("mate.academy.internetshop");
-    private UserService userService
-            = (UserService) INJECTOR.getInstance(UserService.class);
+    private UserService userService = (UserService) INJECTOR.getInstance(UserService.class);
+    private OrderService orderService = (OrderService) INJECTOR.getInstance(OrderService.class);
 
     @Override
     protected void doGet(HttpServletRequest req, HttpServletResponse resp)
             throws ServletException, IOException {
-
-        List<User> userList = userService.getAll();
-        req.setAttribute("users", userList);
-        req.getRequestDispatcher("/WEB-INF/jsp/allUsers.jsp").forward(req, resp);
+        Long userId = (Long) req.getSession().getAttribute("user_id");
+        User user = userService.get(userId);
+        req.setAttribute("order", orderService.getUserOrders(user));
+        req.getRequestDispatcher("/WEB-INF/jsp/userOrders.jsp").forward(req, resp);
     }
 }
