+++ src/main/java/mate/academy/internetshop/controllers/product/CardAddProductController.java
@@ -10,8 +10,8 @@ import mate.academy.internetshop.lib.Injector;
 import mate.academy.internetshop.service.ProductService;
 import mate.academy.internetshop.service.ShoppingCartService;
 
-@WebServlet("/ProductAddToBucket")
-public class AddProductToBucketController extends HttpServlet {
+@WebServlet("/product/add/card")
+public class CardAddProductController extends HttpServlet {
     private static final Injector INJECTOR = Injector.getInstance("mate.academy.internetshop");
     private ShoppingCartService shoppingCartService
             = (ShoppingCartService) INJECTOR.getInstance(ShoppingCartService.class);
@@ -23,7 +23,7 @@ public class AddProductToBucketController extends HttpServlet {
             throws ServletException, IOException {
         Long userId = (Long) req.getSession().getAttribute("user_id");
         shoppingCartService.addProduct(shoppingCartService.getByUserId(userId),
-                productService.get(Long.parseLong(req.getParameter("productId"))));
-        resp.sendRedirect(req.getContextPath() + "/productList");
+                productService.get(Long.parseLong(req.getParameter("product_id"))));
+        resp.sendRedirect(req.getContextPath() + "/product/list");
     }
 }
