+++ src/main/java/mate/academy/internetshop/controllers/product/CardDeleteProductController.java
@@ -1,4 +1,4 @@
-package mate.academy.internetshop.controllers;
+package mate.academy.internetshop.controllers.product;
 
 import java.io.IOException;
 import javax.servlet.ServletException;
@@ -7,23 +7,24 @@ import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import mate.academy.internetshop.lib.Injector;
-import mate.academy.internetshop.model.ShoppingCart;
+import mate.academy.internetshop.model.Product;
+import mate.academy.internetshop.service.ProductService;
 import mate.academy.internetshop.service.ShoppingCartService;
 
-@WebServlet("/shoppingCartShow")
-public class GetUserCartController extends HttpServlet {
+@WebServlet("/cart/delete/product")
+public class CardDeleteProductController extends HttpServlet {
     private static final Injector INJECTOR = Injector.getInstance("mate.academy.internetshop");
     private ShoppingCartService shoppingCartService
-            = (ShoppingCartService) INJECTOR.getInstance(ShoppingCartService.class);
+            = (ShoppingCartService)INJECTOR.getInstance(ShoppingCartService.class);
+    private ProductService productService
+            = (ProductService)INJECTOR.getInstance(ProductService.class);
 
     @Override
     protected void doGet(HttpServletRequest req, HttpServletResponse resp)
             throws ServletException, IOException {
         Long userId = (Long) req.getSession().getAttribute("user_id");
-        ShoppingCart shoppingCart = shoppingCartService.getByUserId(userId);
-        req.setAttribute("shoppingCartList",
-                shoppingCart.getProducts());
-        req.setAttribute("price", shoppingCartService.getTotalPrice(shoppingCart));
-        req.getRequestDispatcher("/WEB-INF/jsp/shoppingCartShow.jsp").forward(req, resp);
+        Product product = productService.get(Long.parseLong(req.getParameter("product_id")));
+        shoppingCartService.deleteProduct(shoppingCartService.getByUserId(userId), product);
+        resp.sendRedirect(req.getContextPath() + "/user/card");
     }
 }
