+++ src/main/java/mate/academy/internetshop/controllers/product/ProductCreateController.java
@@ -9,22 +9,25 @@ import javax.servlet.http.HttpServletResponse;
 import mate.academy.internetshop.lib.Injector;
 import mate.academy.internetshop.model.Product;
 import mate.academy.internetshop.service.ProductService;
-import mate.academy.internetshop.service.ShoppingCartService;
 
-@WebServlet("/productDeleteFromBucket")
-public class DeleteProductFromBucketController extends HttpServlet {
+@WebServlet("/product/create")
+public class ProductCreateController extends HttpServlet {
     private static final Injector INJECTOR = Injector.getInstance("mate.academy.internetshop");
-    private ShoppingCartService shoppingCartService
-            = (ShoppingCartService)INJECTOR.getInstance(ShoppingCartService.class);
     private ProductService productService
-            = (ProductService)INJECTOR.getInstance(ProductService.class);
+            = (ProductService) INJECTOR.getInstance(ProductService.class);
 
     @Override
     protected void doGet(HttpServletRequest req, HttpServletResponse resp)
             throws ServletException, IOException {
-        Long userId = (Long) req.getSession().getAttribute("user_id");
-        Product product = productService.get(Long.parseLong(req.getParameter("productId")));
-        shoppingCartService.deleteProduct(shoppingCartService.getByUserId(userId), product);
-        resp.sendRedirect(req.getContextPath() + "/shoppingCartShow");
+        req.getRequestDispatcher("/WEB-INF/jsp/productCreate.jsp").forward(req,resp);
+    }
+
+    @Override
+    protected void doPost(HttpServletRequest req, HttpServletResponse resp)
+            throws ServletException, IOException {
+        String name = req.getParameter("name");
+        Double price = Double.parseDouble(req.getParameter("price"));
+        productService.create(new Product(name, price));
+        req.getRequestDispatcher("/WEB-INF/jsp/productCreate.jsp").forward(req,resp);
     }
 }
