+++ src/main/java/mate/academy/internetshop/controllers/product/ProductDeleteController.java
@@ -1,27 +1,24 @@
 package mate.academy.internetshop.controllers.product;
 
 import java.io.IOException;
-import java.util.List;
-import javax.servlet.ServletException;
 import javax.servlet.annotation.WebServlet;
 import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import mate.academy.internetshop.lib.Injector;
-import mate.academy.internetshop.model.Product;
 import mate.academy.internetshop.service.ProductService;
 
-@WebServlet("/productList")
-public class GetAllProducts extends HttpServlet {
+@WebServlet("/product/delete")
+public class ProductDeleteController extends HttpServlet {
     private static final Injector INJECTOR = Injector.getInstance("mate.academy.internetshop");
     private ProductService productService
-            = (ProductService) INJECTOR.getInstance(ProductService.class);
+            = (ProductService)INJECTOR.getInstance(ProductService.class);
 
     @Override
     protected void doGet(HttpServletRequest req, HttpServletResponse resp)
-            throws ServletException, IOException {
-        List<Product> productList = productService.getAll();
-        req.setAttribute("products", productList);
-        req.getRequestDispatcher("/WEB-INF/jsp/productList.jsp").forward(req, resp);
+            throws IOException {
+        String id = req.getParameter("product_id");
+        productService.delete(Long.parseLong(id));
+        resp.sendRedirect(req.getContextPath() + "/product/list/admin");
     }
 }
