+++ src/main/java/mate/academy/internetshop/controllers/product/ProductGetListAdminController.java
@@ -11,8 +11,8 @@ import mate.academy.internetshop.lib.Injector;
 import mate.academy.internetshop.model.Product;
 import mate.academy.internetshop.service.ProductService;
 
-@WebServlet("/getProductAdmin")
-public class GetProductAdminController extends HttpServlet {
+@WebServlet("/product/list/admin")
+public class ProductGetListAdminController extends HttpServlet {
     private static final Injector INJECTOR = Injector.getInstance("mate.academy.internetshop");
     private ProductService productService
             = (ProductService)INJECTOR.getInstance(ProductService.class);
