+++ src/main/java/mate/academy/internetshop/controllers/product/ProductGetListController.java
@@ -1,24 +1,27 @@
 package mate.academy.internetshop.controllers.product;
 
 import java.io.IOException;
+import java.util.List;
+import javax.servlet.ServletException;
 import javax.servlet.annotation.WebServlet;
 import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import mate.academy.internetshop.lib.Injector;
+import mate.academy.internetshop.model.Product;
 import mate.academy.internetshop.service.ProductService;
 
-@WebServlet("/productDelete")
-public class DeleteProductController extends HttpServlet {
+@WebServlet("/product/list")
+public class ProductGetListController extends HttpServlet {
     private static final Injector INJECTOR = Injector.getInstance("mate.academy.internetshop");
     private ProductService productService
-            = (ProductService)INJECTOR.getInstance(ProductService.class);
+            = (ProductService) INJECTOR.getInstance(ProductService.class);
 
     @Override
     protected void doGet(HttpServletRequest req, HttpServletResponse resp)
-            throws IOException {
-        String id = req.getParameter("productId");
-        productService.delete(Long.parseLong(id));
-        resp.sendRedirect(req.getContextPath() + "/getProductAdmin");
+            throws ServletException, IOException {
+        List<Product> productList = productService.getAll();
+        req.setAttribute("products", productList);
+        req.getRequestDispatcher("/WEB-INF/jsp/productList.jsp").forward(req, resp);
     }
 }
